requirements.txt
Flask==3.0.3
pymongo==4.10.1
python-dotenv==1.0.1
Requests==2.32.3
tweepy==4.14.0
Werkzeug==3.1.3
